refactor(llms): rename llms module to train
