+++ src/_mnemonic_enhancer/mnemonic_ft.py
@@ -12,7 +12,7 @@ from pydantic import ValidationError
 from structlog import getLogger
 
 from src.data.data_io import read_csv_file, write_jsonl_file
-from src.llms.openai import (
+from src.train.openai import (
     upload_file_to_openai,
     validate_openai_file,
 )
