--- src/sft/__init__.py
@@ -1,1 +1,0 @@
-"""Module for supervised fine-tuning (SFT) of language models hosted on Hugging Face Hub."""
